backend/requirement.txt
fastapi
uvicorn
pydantic
numpy
scikit-learn
joblib
torch
transformers
fastapi
uvicorn
pymupdf
pillow
google-generativeai
fastapi
uvicorn
pymongo
motor
python-multipart
numpy
pandas
scikit-learn
tensorflow
torch
transformers
openai
opencv-python
geopy
requests
pydantic
python-dotenv
pytesseract
passLib